+++ internal/aws/config.go
@@ -1,4 +1,4 @@
-package config
+package aws
 
 import (
 	"fmt"
