+++ internal/aws/s3.go
@@ -1,4 +1,4 @@
-package s3
+package aws
 
 import (
 	"github.com/aws/aws-sdk-go-v2/aws"
